Removes X3DNA dependency
