+++ open3SPN2/DNA_atomic.csv
@@ -1,0 +1,132 @@
+name,base,x,y,z
+P,A,-0.299,9.399,-1.529
+O1P,A,-0.377,10.734,-2.162
+O2P,A,0.714,9.245,-0.460
+O5',A,-1.738,8.985,-0.968
+C5',A,-2.674,8.343,-1.855
+C4',A,-3.346,7.182,-1.148
+O4',A,-2.596,5.941,-1.284
+C3',A,-3.530,7.338,0.361
+O3',A,-4.771,6.752,0.737
+C2',A,-2.398,6.514,0.976
+C1',A,-2.479,5.346,0.000
+N9,A,-1.291,4.498,0.000
+C8,A,0.024,4.897,0.000
+N7,A,0.877,3.902,0.000
+C5,A,0.071,2.771,0.000
+C6,A,0.369,1.398,0.000
+N6,A,1.611,0.909,0.000
+N1,A,-0.668,0.532,0.000
+C2,A,-1.912,1.023,0.000
+N3,A,-2.320,2.290,0.000
+C4,A,-1.267,3.124,0.000
+P,C,-0.275,9.443,-1.528
+O1P,C,-0.347,10.779,-2.161
+O2P,C,0.736,9.284,-0.459
+O5',C,-1.717,9.036,-0.967
+C5',C,-2.657,8.399,-1.854
+C4',C,-3.334,7.242,-1.148
+O4',C,-2.590,5.997,-1.284
+C3',C,-3.517,7.399,0.362
+O3',C,-4.762,6.819,0.737
+C2',C,-2.390,6.569,0.977
+C1',C,-2.477,5.402,0.000
+N1,C,-1.285,4.542,0.000
+C2,C,-1.472,3.158,0.000
+O2,C,-2.628,2.709,0.001
+N3,C,-0.391,2.344,0.000
+C4,C,0.837,2.868,0.000
+N4,C,1.875,2.027,0.001
+C5,C,1.056,4.275,0.000
+C6,C,-0.023,5.068,0.000
+P,G,-0.296,9.452,-1.527
+O1P,G,-0.375,10.788,-2.160
+O2P,G,0.716,9.298,-0.458
+O5',G,-1.736,9.038,-0.967
+C5',G,-2.672,8.396,-1.854
+C4',G,-3.344,7.235,-1.148
+O4',G,-2.593,5.994,-1.284
+C3',G,-3.528,7.391,0.362
+O3',G,-4.769,6.805,0.737
+C2',G,-2.396,6.566,0.977
+C1',G,-2.477,5.399,0.000
+N9,G,-1.289,4.551,0.000
+C8,G,0.023,4.962,0.000
+N7,G,0.870,3.969,0.000
+C5,G,0.071,2.833,0.000
+C6,G,0.424,1.460,0.000
+O6,G,1.554,0.955,0.000
+N1,G,-0.700,0.641,0.000
+C2,G,-1.999,1.087,0.000
+N2,G,-2.949,0.139,-0.001
+N3,G,-2.342,2.364,0.001
+C4,G,-1.265,3.177,0.000
+P,I,-0.296,9.452,-1.527
+O1P,I,-0.375,10.788,-2.160
+O2P,I,0.716,9.298,-0.458
+O5',I,-1.736,9.038,-0.967
+C5',I,-2.672,8.396,-1.854
+C4',I,-3.344,7.235,-1.148
+O4',I,-2.593,5.994,-1.284
+C3',I,-3.528,7.391,0.362
+O3',I,-4.769,6.805,0.737
+C2',I,-2.396,6.566,0.977
+C1',I,-2.477,5.399,0.000
+N9,I,-1.289,4.551,0.000
+C8,I,0.023,4.962,0.000
+N7,I,0.870,3.969,0.000
+C5,I,0.071,2.833,0.000
+C6,I,0.424,1.460,0.000
+O6,I,1.554,0.955,0.000
+N1,I,-0.700,0.641,0.000
+C2,I,-1.999,1.087,0.000
+N3,I,-2.342,2.364,0.001
+C4,I,-1.265,3.177,0.000
+C1',PSU,-2.506,5.371,0.000
+N1,PSU,1.087,4.295,0.000
+C2,PSU,1.037,2.915,0.000
+O2,PSU,2.036,2.217,0.000
+N3,PSU,-0.229,2.383,0.000
+C4,PSU,-1.422,3.076,0.000
+O4,PSU,-2.485,2.453,0.000
+C5,PSU,-1.284,4.500,0.000
+C6,PSU,-0.064,5.048,0.000
+P,T,-0.300,9.407,-1.526
+O1P,T,-0.379,10.743,-2.159
+O2P,T,0.712,9.253,-0.458
+O5',T,-1.740,8.993,-0.966
+C5',T,-2.676,8.351,-1.853
+C4',T,-3.348,7.190,-1.147
+O4',T,-2.597,5.949,-1.283
+C3',T,-3.532,7.346,0.362
+O3',T,-4.773,6.760,0.737
+C2',T,-2.400,6.521,0.977
+C1',T,-2.481,5.354,0.000
+N1,T,-1.284,4.500,0.000
+C2,T,-1.462,3.135,0.000
+O2,T,-2.562,2.608,0.000
+N3,T,-0.298,2.407,0.000
+C4,T,0.994,2.897,0.000
+O4,T,1.944,2.119,0.000
+C5,T,1.106,4.338,0.000
+C5M,T,2.466,4.961,0.001
+C6,T,-0.024,5.057,0.000
+P,U,-0.300,9.408,-1.525
+O1P,U,-0.378,10.743,-2.158
+O2P,U,0.712,9.253,-0.456
+O5',U,-1.740,8.993,-0.965
+C5',U,-2.675,8.351,-1.853
+C4',U,-3.347,7.190,-1.147
+O4',U,-2.597,5.949,-1.283
+C3',U,-3.532,7.346,0.362
+O3',U,-4.773,6.759,0.737
+C2',U,-2.400,6.521,0.978
+C1',U,-2.481,5.354,0.000
+N1,U,-1.284,4.500,0.000
+C2,U,-1.462,3.131,0.000
+O2,U,-2.563,2.608,0.000
+N3,U,-0.302,2.397,0.000
+C4,U,0.989,2.884,0.000
+O4,U,1.935,2.094,-0.001
+C5,U,1.089,4.311,0.000
+C6,U,-0.024,5.053,0.000
